Move common logic to FunderMaps.AspNetCore
