+++ src/FunderMaps.AspNetCore/Controllers/UserController.cs
@@ -1,15 +1,14 @@
-﻿using AutoMapper;
-using FunderMaps.Controllers;
+using AutoMapper;
+using FunderMaps.AspNetCore.DataTransferObjects;
 using FunderMaps.Core.Authentication;
 using FunderMaps.Core.Entities;
 using FunderMaps.Core.Managers;
-using FunderMaps.WebApi.DataTransferObjects;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
 
 #pragma warning disable CA1062 // Validate arguments of public methods
-namespace FunderMaps.WebApi.Controllers.Application
+namespace FunderMaps.AspNetCore.Controllers
 {
     /// <summary>
     ///     Endpoint controller for user operations.
@@ -19,7 +18,7 @@ namespace FunderMaps.WebApi.Controllers.Application
     ///     user session. Therefore the user context must be active.
     /// </remarks>
     [Route("user")]
-    public class UserController : BaseApiController
+    public class UserController : ControllerBase
     {
         private readonly IMapper _mapper;
         private readonly AuthManager _authManager;
