+++ tests/FunderMaps.IntegrationTests/Backend/Application/OrganizationSetupTests.cs
@@ -1,5 +1,5 @@
-﻿using FunderMaps.Testing.Faker;
-using FunderMaps.WebApi.DataTransferObjects;
+﻿using FunderMaps.AspNetCore.DataTransferObjects;
+using FunderMaps.Testing.Faker;
 using System.Net;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
