+++ tests/FunderMaps.IntegrationTests/Backend/Application/OrganizationTests.cs
@@ -1,6 +1,6 @@
-﻿using FunderMaps.Core.Types;
+﻿using FunderMaps.AspNetCore.DataTransferObjects;
+using FunderMaps.Core.Types;
 using FunderMaps.Testing.Faker;
-using FunderMaps.WebApi.DataTransferObjects;
 using System.Net;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
